httpio: move to gridstatus.httpio module
